+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 2.5.3
 * Update API to 5.9.0
 * Fix npm vulnerabilities
-* Refactor code
+* Refactor event handlers into behavior.ts
 
 ## 2.5.2
 * Fix backwards compatibility with 2.4.0